refector: utils함수 분리
