+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './timeAgo';
 export * from './timeFormat';
 export * from './numberWithCommas';
 export * from './timeFormatWithoutTime';
+export * from './media';
